# TaskerSwagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:3133'
basePath: /
...
paths:
...
  /tasks:
    get:
      tags:
        - tasker-controller
      summary: getAllTasks
      operationId: getAllTasksUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/TaskViewModel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - tasker-controller
      summary: createTask
      operationId: createTaskUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: taskViewModel
          description: taskViewModel
          required: true
          schema:
            $ref: '#/definitions/TaskViewModel'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/TaskViewModel'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
        - tasker-controller
      summary: updateTask
      operationId: updateTaskUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: taskViewModel
          description: taskViewModel
          required: true
          schema:
            $ref: '#/definitions/TaskViewModel'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/tasks/category/{category}':
    get:
      tags:
        - tasker-controller
      summary: getTasksByCategory
      operationId: getTasksByCategoryUsingGET
      produces:
        - '*/*'
      parameters:
        - name: category
          in: path
          description: category
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/TaskViewModel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/tasks/{id}':
    get:
      tags:
        - tasker-controller
      summary: getTaskById
      operationId: getTaskByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/TaskViewModel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - tasker-controller
      summary: deleteTask
      operationId: deleteTaskUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
definitions:
  TaskViewModel:
    type: object
    properties:
      advertisement:
        type: string
      category:
        type: string
      createDate:
        type: string
        example: yyyy-MM-dd
      description:
        type: string
      dueDate:
        type: string
        example: yyyy-MM-dd
      id:
        type: integer
        format: int32
    title: TaskViewModel